+++ wpull/app_test.py
@@ -118,6 +118,7 @@ class TestApp(GoodAppTestCase):
             '--progress', 'dot',
             '--secure-protocol', 'TLSv1',
             '--convert-links', '--backup-converted',
+            '--accept', '*',
         ])
         with cd_tempdir():
             engine = Builder(args).build()